requirements.txt
Django==1.9
django-pure-pagination==0.3.0
MySQL-python==1.2.5
Pillow==3.4.2
